pyproject.toml
[tool]
[tool.poetry]
name = "humlab-westac"
version = "0.2.11"
description = "Welfare State Analytics"
authors = ["Roger Mähler <roger.mahler@hotmail.com>"]
packages = [
    { include = "notebooks" }
]
classifiers = [
    'License :: OSI Approved :: Apache Software License',
    'Operating System :: OS Independent',
    'Topic :: Software Development',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://westac.se"
repository = "https://github.com/welfare-state-analytics/welfare_state_analytics"

[tool.poetry.dependencies]
alive-progress = "^1.6.1"
appnope = "*"
attrs = "*"
bokeh = "*"
click = "*"
colorcet = "*"
cython = "*"
cytoolz = "*"
decorator = "*"
ftfy = "*"
gensim = "*"
holoviews = "*"
humanfriendly = "*"
ipyaggrid = "==0.2.1"
ipykernel = "*"
ipympl = "*"
ipywidgets = "==7.5.1"
joblib = "*"
jupyter-bokeh = "*"
jupyterlab = "==2.2.9"
lxml = "*"
markdown = "*"
matplotlib = "*"
memoization = "*"
more-itertools = "*"
networkx = "*"
nltk = "*"
numpy = "*"
openpyxl = "*"
pandas = "*"
pandas-bokeh = "*"
py4j = "*"
pydotplus = "*"
pyemd = "*"
python = "==3.8.5"
python-dotenv = "*"
python-louvain = "*"
qgrid = "^1.3.1"
scattertext = "*"
scikit-learn = "*"
scipy = "*"
seaborn = "*"
sklearn = "*"
spacy = "==2.3.2"
statsmodels = "*"
textacy = "==0.10.1"
tqdm = "*"
traitlets = "==4.3.3"
wordcloud = "*"
xlrd = "*"
xmltodict = "*"
ipyfilechooser = "^0.4.0"

cutter-ng = "^2.5"
matplotlib-venn = "^0.11.6"
ipycytoscape = "1.1.0"
loguru = "^0.5.3"
pyarrow = "3.0.0"
ipyregulartable = "^0.2.0"
perspective-python = "0.6.2"
humlab-penelope = "^0.3.13"

[tool.poetry.dev-dependencies]
black = "==20.*,>=20.8.0.b1"
coverage = "*"
dephell = "*"
flake8 = "==3.*,>=3.8.4"
flake8-black = "==0.*,>=0.2.1"
flake8-pytest-style = "==1.*,>=1.3.0"
invoke = "*"
isort = "==5.*,>=5.5.4"
mypy = "==0.*,>=0.782.0"
nbstripout = "*"
pep8-naming = "==0.*,>=0.11.1"
pre-commit = "*"
pycodestyle = "==2.*,>=2.6.0"
pyflakes = "==2.*,>=2.2.0"
pylint = "*"
pytest = "*"
pytest-cov = "==2.*,>=2.10.1"
rope = "*"
yapf = "==0.*,>=0.30.0"
debugpy = "==1.*,>=1.2.1"
jupytext = "==1.7.1"
pyinstrument = "^3.4.1"

[tool.poetry.scripts]
# vectorize_corpus = "penelope.scripts.vectorize_corpus:main"
# compute_topic_model = "penelope.scripts.compute_topic_model:main"
# concept_cooccurrence = "penelope.scripts.concept_co_occurrence:main"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-ra",
    "-q",
    # "-s",
]
log_cli = "False"
testpaths = [
    "tests",
]
norecursedirs = ['notebooks', '.git', 'data', 'output', 'deprecated', '__pycache__', 'tmp']
python_files = "*_test.py"
python_functions = "test_*"
filterwarnings = [
#   "ignore",
  "ignore::_pytest.warning_types.PytestCollectionWarning",
  "ignore::DeprecationWarning",
]

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "poetry", path = "pyproject.toml"}

[tool.black]
line-length = 120
target-version = ["py38"]
skip-string-normalization = true
# include = '\.pyi?$'
exclude = '''
(
  \.git
  | \.pypy_cache
  | \.pytest_cache
  | \.ipynb_checkpoints
  | deprecated
)
'''

[tool.isort]
# https://github.com/timothycrosley/isort/wiki/isort-Settings
profile = "black"
line_length = 120
py_version = "auto"
float_to_top = true
src_paths = [ "notebooks", "westac", "tests" ]
known_first_party = [ "__paths__" ]

[tool.coverage]
# https://github.com/nedbat/coveragepy

    [tool.coverage.run]
    source = ["notebooks", "westac"]
    branch = true
    omit = [
        '**/__init__.py',
        '**/__paths__.py',
        'notebooks/legacy/**/*.py',
        'notebooks/co_occurrence/co_occurrence_trends.py',
        'notebooks/pos_statistics/pos_statistics.py',
        'notebooks/word_trends/word_trends.py'
    ]

    [tool.coverage.report]
    exclude_lines = [
        "pragma: no cover",
        "raise NotImplementedError"
    ]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
